Add an admission webhook to check cluster ID (kube-system UID) before updating
